Im VMInterface Skelette für zwei einfache Befehl hinzugefügt und klasse RepositoryViewModelAPI erstellt
