+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/Button.kt
@@ -1,0 +1,17 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+class Button(
+    override val id: Long,
+    val name: String,
+    val columnId: Long,
+    val value: Int
+): Identifiable {
+    override fun deleteIsPossible(): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    //@throws IllegalOperationException
+    override fun delete() {
+        TODO("Not yet implemented")
+    }
+}
