+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/DataType.kt
@@ -1,0 +1,9 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+enum class DataType {
+    WHOLE_NUMBER,
+    FLOATING_POINT_NUMBER,
+    TIME,
+    STRING
+    //Todo Fields
+}
