+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/DotSize.kt
@@ -1,0 +1,8 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+enum class DotSize {
+    SMALL,
+    MEDIUM,
+    BIG
+    //Todo Fields
+}
