+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/Graph.kt
@@ -1,0 +1,8 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+import android.graphics.drawable.Drawable
+
+interface Graph: Identifiable{
+    override val id: Long
+    val image: Drawable
+}
