+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/GraphTemplate.kt
@@ -1,0 +1,5 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+class GraphTemplate {
+    //TODO Implementierung
+}
