+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/Identifiable.kt
@@ -1,0 +1,8 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+interface Identifiable {
+    val id: Long
+    fun deleteIsPossible(): Boolean
+    //@throws IllegalOperationException
+    fun delete()
+}
