+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/LineType.kt
@@ -1,0 +1,8 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+enum class LineType {
+    DOTTED,
+    NONE,
+    CONTINUOUS
+    //Todo Fields
+}
