+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/Notification.kt
@@ -1,0 +1,18 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+import java.time.LocalDateTime
+
+class Notification(
+    override val id: Long,
+    val message: String,
+    val time: LocalDateTime
+): Identifiable {
+    override fun deleteIsPossible(): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    //@throws IllegalOperationException
+    override fun delete() {
+        TODO("Not yet implemented")
+    }
+}
