+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/PieChart.kt
@@ -1,0 +1,21 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+import android.graphics.Color
+import android.graphics.drawable.Drawable
+
+class PieChart(
+    override val id: Long,
+    override val image: Drawable,
+    val color: List<Color>,
+    val mapping: List<Column>,
+    val showPercentages: Boolean
+): Graph {
+    override fun deleteIsPossible(): Boolean {
+        TODO("Not yet implemented")
+    }
+
+    //@throws IllegalOperationException
+    override fun delete() {
+        TODO("Not yet implemented")
+    }
+}
