+++ DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelAPI/communicationClasses/Post.kt
@@ -1,0 +1,5 @@
+package com.pseandroid2.dailydata.repository.viewModelAPI.communicationClasses
+
+class Post {
+    //Todo implementierung
+}
