--- DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelInterface/communicationClasses/DataType.kt
@@ -1,9 +1,0 @@
-package com.pseandroid2.dailydata.repository.viewModelInterface.communicationClasses
-
-enum class DataType {
-    WHOLE_NUMBER,
-    FLOATING_POINT_NUMBER,
-    TIME,
-    STRING
-    //Todo Fields
-}
