--- DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelInterface/communicationClasses/DotSize.kt
@@ -1,8 +1,0 @@
-package com.pseandroid2.dailydata.repository.viewModelInterface.communicationClasses
-
-enum class DotSize {
-    SMALL,
-    MEDIUM,
-    BIG
-    //Todo Fields
-}
