--- DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelInterface/communicationClasses/Graph.kt
@@ -1,8 +1,0 @@
-package com.pseandroid2.dailydata.repository.viewModelInterface.communicationClasses
-
-import android.graphics.drawable.Drawable
-
-interface Graph: Identifiable{
-    override val id: Long
-    val image: Drawable
-}
