--- DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelInterface/communicationClasses/GraphTemplate.kt
@@ -1,5 +1,0 @@
-package com.pseandroid2.dailydata.repository.viewModelInterface.communicationClasses
-
-class GraphTemplate {
-    //TODO Implementierung
-}
