--- DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelInterface/communicationClasses/Identifiable.kt
@@ -1,8 +1,0 @@
-package com.pseandroid2.dailydata.repository.viewModelInterface.communicationClasses
-
-interface Identifiable {
-    val id: Long
-    fun deleteIsPossible(): Boolean
-    //@throws IllegalOperationException
-    fun delete()
-}
