--- DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelInterface/communicationClasses/LineType.kt
@@ -1,8 +1,0 @@
-package com.pseandroid2.dailydata.repository.viewModelInterface.communicationClasses
-
-enum class LineType {
-    DOTTED,
-    NONE,
-    CONTINUOUS
-    //Todo Fields
-}
