--- DailyData/app/src/main/java/com/pseandroid2/dailydata/repository/viewModelInterface/communicationClasses/ProjectTemplate.kt
@@ -1,15 +1,0 @@
-package com.pseandroid2.dailydata.repository.viewModelInterface.communicationClasses
-
-class ProjectTemplate : Identifiable{
-    override val id: Long
-        get() = TODO("Not yet implemented")
-
-    override fun deleteIsPossible(): Boolean {
-        TODO("Not yet implemented")
-    }
-    //@throws IllegalOperationException
-    override fun delete() {
-        TODO("Not yet implemented")
-    }
-    //TODO Implementierung
-}
